requirements.txt
amqp==5.1.1
anyascii==0.3.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==4.1.0
bleach==6.0.0
celery==5.3.0
certifi==2023.5.7
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.9
django-cogwheels==0.3
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-filter==22.1
django-jalali-date==1.0.2
django-markdownx==4.0.2
django-minify-html==1.6.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-quill-editor==0.1.40
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
importlib-metadata==6.7.0
jdatetime==4.1.1
kombu==5.3.1
l18n==2021.3
Markdown==3.4.3
md-mermaid==0.1.1
medianasms==1.1.0
minify_html==0.11.1
mysqlclient==2.1.1
openpyxl==3.1.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.25.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3
tinycss2==1.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wagtail==4.2.4
wagtailmenus==3.1.8
wcwidth==0.2.6
webencodings==0.5.1
Willow==1.4.1
zipp==3.15.0
